Add files via upload
This project is designed to give me a hands-on experience with supervised learning and 
work towards developing conclusions for a potential client on a real-world dataset. Many companies 
today collect vast amounts of data on customers and clients, and have a strong desire to understand 
the meaningful relationships hidden in their customer base.